requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
fastapi==0.110.1
frozenlist==1.4.1
google-auth==2.23.0
google-auth-oauthlib==0.4.6
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
iso8601==1.1.0
multidict==6.0.5
oauthlib==3.2.2
packaging==24.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
starlette==0.37.2
tortoise-orm==0.20.0
typing_extensions==4.11.0
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
